backend/requirements.txt
Flask
flask-cors
flask_limiter
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
python-dotenv